+++ scripts/migration/ths_r4_sanity.py
@@ -25,7 +25,7 @@ logging.getLogger('toshi_hazard_store').setLevel(logging.WARNING)
 
 from nzshm_common import location
 from nzshm_common.grids import load_grid
-from nzshm_common.location.code_location import CodedLocation
+from nzshm_common.location.coded_location import CodedLocation
 from pynamodb.models import Model
 
 import toshi_hazard_store  # noqa: E402