+++ scripts/ths_cache.py
@@ -10,7 +10,7 @@ import time
 import click
 import pandas as pd
 from nzshm_common.grids import RegionGrid, load_grid
-from nzshm_common.location.code_location import CodedLocation
+from nzshm_common.location.coded_location import CodedLocation
 from nzshm_common.location.location import LOCATION_LISTS, LOCATIONS, location_by_id
 
 from toshi_hazard_store import model, query