+++ tests/test_query_disaggs.py
@@ -3,7 +3,7 @@ import unittest
 
 import numpy as np
 from moto import mock_dynamodb
-from nzshm_common.location.code_location import CodedLocation
+from nzshm_common.location.coded_location import CodedLocation
 from nzshm_common.location.location import LOCATIONS_BY_ID
 
 from toshi_hazard_store import model, query