+++ tests/test_query_hazard_caching.py
@@ -4,7 +4,7 @@ import unittest
 from importlib import reload
 
 from moto import mock_dynamodb
-from nzshm_common.location.code_location import CodedLocation
+from nzshm_common.location.coded_location import CodedLocation
 from nzshm_common.location.location import LOCATIONS_BY_ID
 from pynamodb.models import Model