+++ toshi_hazard_store/model/revision_4/hazard_models.py
@@ -2,7 +2,7 @@
 
 import logging
 
-from nzshm_common.location.code_location import CodedLocation
+from nzshm_common.location.coded_location import CodedLocation
 from pynamodb.attributes import ListAttribute, NumberAttribute, UnicodeAttribute, VersionAttribute
 from pynamodb.models import Model
 from pynamodb_attributes import TimestampAttribute