+++ toshi_hazard_store/query/disagg_queries.py
@@ -14,7 +14,7 @@ import itertools
 import logging
 from typing import Iterable, Iterator, List, Type, Union
 
-from nzshm_common.location.code_location import CodedLocation
+from nzshm_common.location.coded_location import CodedLocation
 from pynamodb.expressions.condition import Condition
 
 from toshi_hazard_store.model import (